pythonProject1/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-limiter==3.5.0
flask-talisman==1.1.0
pyjwt==2.8.0
psycopg2==2.9.9
python-dotenv==1.0.0
email-validator==2.0.0
openpyxl==2.0.3
pandas==3.1.5
